+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 * Fixing `xan enum -A/--accumulate`.
 * Fixing `xan from -f tar` when tarball archive is not gzipped.
 * Fixing `min` & `max` moonblade function when passing a list of numbers.
+* Fixing `xan flatten -H` edge cases.
 
 *Performance*